Now only author can delete their blogs and comments in their blogs, add cors handling

+++ controllers/CommentController.js
@@ -1,4 +1,5 @@
 const Comment = require("../ models/Comment");
+const Blog = require("../ models/Blog");
 
 let postComment = (req, res) => {
   const comment = new Comment({
@@ -16,13 +17,20 @@ let postComment = (req, res) => {
 };
 
 let deleteComment = (req, res) => {
-  Comment.findByIdAndDelete(req.params.commentId)
-    .then((data) => {
-      res.redirect("/blogs/detail/" + req.params.blogId);
-    })
-    .catch((err) => {
-      console.log(err);
-    });
+  if (req.isAuthenticated()) {
+    Blog.findById(req.params.blogId)
+      .then((blog) => {
+        if (req.user.username === blog.author) {
+          return Comment.findByIdAndDelete(req.params.commentId);
+        } else res.redirect("/blogs/detail/" + req.params.blogId);
+      })
+      .then((result) => {
+        res.redirect("/blogs/detail/" + req.params.blogId);
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  } else res.redirect("/blogs/detail/" + req.params.blogId);
 };
 
 module.exports = {
